Move Drive view code to separate page
