+++ frontend/src/pages/LandingPage.tsx
@@ -1,0 +1,5 @@
+"use client"
+
+export default function LandingPage() {
+  return <div className="p-4 text-center text-white">Landing Page Placeholder</div>
+}
