+++ frontend/src/pages/SignIn.tsx
@@ -1,0 +1,5 @@
+"use client"
+
+export default function SignIn() {
+  return <div className="p-4 text-center text-white">Sign In Placeholder</div>
+}
